graphql_service/requirements.txt
graphql-core==3.2.5
httpx==0.28.1
python-dotenv==1.0.1
requests==2.32.3
starlette==0.41.3
strawberry-graphql==0.254.0
uvicorn==0.32.1
fastapi
